refactor: AI기능와 메시지 기능을 분리
